Command is the only symbol that can have multiple Children
